SOFTWARE-647: Remove test user only when tests created it.  This change entailed several other small changes:
* Deprecated the "--cleanup" option and added a warning message about it
* Added a "--no-cleanup" option
* Made the "special_user" module always run
* Made it safe to always run the "special_user" module
* Improved the "special_user" tests to check for failures more clearly
* Cleaned up the "special_user" module a bit
* Remove the test user only when we know we added it
* Made the append_file() method safe for non-existent files


git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-test/trunk@14713 4e558342-562e-0410-864c-e07659590f8c

+++ osgtest/tests/special_cleanup.py
@@ -67,15 +67,20 @@ class TestCleanup(unittest.TestCase):
         command = ['rpm', '--quiet', '--erase'] + rpm_erase_list
         core.check_system(command, 'Remove %d packages' % (package_count))
 
-    def test_02_remove_test_user(self):
+    def test_02_restore_mapfile(self):
         files.restore(core.config['system.mapfile'])
+
+    def test_03_remove_test_user(self):
+        if not core.state['general.user_added']:
+            core.skip('did not add user')
+            return
 
         username = core.options.username
         password_entry = pwd.getpwnam(username)
         globus_dir = os.path.join(password_entry.pw_dir, '.globus')
 
         command = ('userdel', username)
-        core.check_system(command, 'Remove user %s' % (username))
+        core.check_system(command, "Remove user '%s'" % (username))
 
         files.remove(os.path.join(globus_dir, 'usercert.pem'))
         files.remove(os.path.join(globus_dir, 'userkey.pem'))
